+++ tests/workers/rollout/test_sglang_async_rollout_mcp_tools.py
@@ -27,6 +27,7 @@ from utils_sglang import get_rollout_config, prepare_inputs
 
 from verl.protocol import DataProto
 from verl.tools.mcp_search_tool import MCPSearchTool
+from verl.tools.utils.mcp_clients.McpClientManager import MCPClientManager
 from verl.workers.rollout.schemas import AsyncRolloutRequest, AsyncRolloutRequestStateEnum, Message
 from verl.workers.rollout.sglang_rollout.sglang_rollout import SGLangRollout
 
@@ -150,10 +151,44 @@ class TestRolloutWithMCPSearchTools:
         prompts = DataProto(batch=prompt_dict, non_tensor_batch={"raw_prompt": messages, "tools_kwargs": tools_kwargs, "index": index})
         return prompts
 
+    @patch.object(MCPClientManager, "fetch_tool_schemas", return_value=None)
     @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
     @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
     @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
-    def test_tools_registration(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config):
+    def test_tools_registration(self, mock_env, mock_engine, mock_sampling, mock_fetch, search_rollout_config, qwen_tokenizer, qwen_model_config):
+        tool_schema = [
+            {
+                "type": "function",
+                "function": {
+                    "name": "tavily_search_tool",
+                    "description": "A powerful web search tool...",
+                    "parameters": {
+                        "type": "object",
+                        "properties": {
+                            "what_is_your_intent": {"type": "string", "description": "Describe your intent for using Tavily"},
+                            "query": {"type": "string", "description": "Search query"},
+                            "search_depth": {"type": "string", "description": "The depth of the search ('basic' or 'advanced')"},
+                            "topic": {"type": "string", "description": "The category of the search ('general' or 'news')"},
+                            "days": {"type": "integer", "description": "Number of days back to include in search results (only for 'news' topic)"},
+                            "time_range": {"type": "string", "description": "Time range for results ('day', 'week', 'month', 'year', 'd', 'w', 'm', 'y')"},
+                            "include_domains": {"type": "array", "description": "List of domains to specifically include in search results"},
+                            "exclude_domains": {"type": "array", "description": "List of domains to specifically exclude from search results"},
+                            "include_answer": {"type": "boolean", "description": "Whether to include an answer summary generated by an LLM"},
+                            "include_raw_content": {"type": "boolean", "description": "Whether to include the cleaned and parsed HTML content of each result"},
+                            "include_images": {"type": "boolean", "description": "Whether to include images from search results"},
+                            "include_image_descriptions": {"type": "boolean", "description": "Whether to include descriptions with images"},
+                            "max_results": {"type": "integer", "description": "Maximum number of results to return (5-20)"},
+                            "async_search": {"type": "boolean", "description": "Whether to perform the search asynchronously"},
+                        },
+                        "required": ["what_is_your_intent", "query"],
+                    },
+                    "strict": False,
+                },
+            }
+        ]
+        # Mock mcp registered tool as a predefined schema for avoiding ci timeout
+        mock_fetch.return_value = tool_schema
+
         rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         assert len(rollout._tool_schemas) != 0
         assert "tavily_search_tool" in rollout._tool_map.keys()
@@ -163,20 +198,86 @@ class TestRolloutWithMCPSearchTools:
         # depend on the tokenizer
         assert rollout._tool_call_parser_type == "qwen25"
 
+    @patch.object(MCPClientManager, "fetch_tool_schemas", return_value=None)
     @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
     @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
     @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
-    def test_rollout_req_creation(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto):
+    def test_rollout_req_creation(self, mock_env, mock_engine, mock_sampling, mock_fetch, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto):
+        tool_schema = [
+            {
+                "type": "function",
+                "function": {
+                    "name": "tavily_search_tool",
+                    "description": "A powerful web search tool...",
+                    "parameters": {
+                        "type": "object",
+                        "properties": {
+                            "what_is_your_intent": {"type": "string", "description": "Describe your intent for using Tavily"},
+                            "query": {"type": "string", "description": "Search query"},
+                            "search_depth": {"type": "string", "description": "The depth of the search ('basic' or 'advanced')"},
+                            "topic": {"type": "string", "description": "The category of the search ('general' or 'news')"},
+                            "days": {"type": "integer", "description": "Number of days back to include in search results (only for 'news' topic)"},
+                            "time_range": {"type": "string", "description": "Time range for results ('day', 'week', 'month', 'year', 'd', 'w', 'm', 'y')"},
+                            "include_domains": {"type": "array", "description": "List of domains to specifically include in search results"},
+                            "exclude_domains": {"type": "array", "description": "List of domains to specifically exclude from search results"},
+                            "include_answer": {"type": "boolean", "description": "Whether to include an answer summary generated by an LLM"},
+                            "include_raw_content": {"type": "boolean", "description": "Whether to include the cleaned and parsed HTML content of each result"},
+                            "include_images": {"type": "boolean", "description": "Whether to include images from search results"},
+                            "include_image_descriptions": {"type": "boolean", "description": "Whether to include descriptions with images"},
+                            "max_results": {"type": "integer", "description": "Maximum number of results to return (5-20)"},
+                            "async_search": {"type": "boolean", "description": "Whether to perform the search asynchronously"},
+                        },
+                        "required": ["what_is_your_intent", "query"],
+                    },
+                    "strict": False,
+                },
+            }
+        ]
+        # Mock mcp registered tool as a predefined schema for avoiding ci timeout
+        mock_fetch.return_value = tool_schema
         rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         req_list = rollout._preprocess_prompt_to_async_rollout_requests(search_data_proto, n=1)
         assert len(req_list) == 1
         assert req_list[0].state == AsyncRolloutRequestStateEnum.PENDING
-        assert len(req_list[0].tools) == 1
+        assert len(req_list[0].tool_schemas) == 1
 
+    @patch.object(MCPClientManager, "fetch_tool_schemas", return_value=None)
     @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
     @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
     @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
-    def test_over_size_case(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
+    def test_over_size_case(self, mock_env, mock_engine, mock_sampling, mock_fetch, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
+        tool_schema = [
+            {
+                "type": "function",
+                "function": {
+                    "name": "tavily_search_tool",
+                    "description": "A powerful web search tool...",
+                    "parameters": {
+                        "type": "object",
+                        "properties": {
+                            "what_is_your_intent": {"type": "string", "description": "Describe your intent for using Tavily"},
+                            "query": {"type": "string", "description": "Search query"},
+                            "search_depth": {"type": "string", "description": "The depth of the search ('basic' or 'advanced')"},
+                            "topic": {"type": "string", "description": "The category of the search ('general' or 'news')"},
+                            "days": {"type": "integer", "description": "Number of days back to include in search results (only for 'news' topic)"},
+                            "time_range": {"type": "string", "description": "Time range for results ('day', 'week', 'month', 'year', 'd', 'w', 'm', 'y')"},
+                            "include_domains": {"type": "array", "description": "List of domains to specifically include in search results"},
+                            "exclude_domains": {"type": "array", "description": "List of domains to specifically exclude from search results"},
+                            "include_answer": {"type": "boolean", "description": "Whether to include an answer summary generated by an LLM"},
+                            "include_raw_content": {"type": "boolean", "description": "Whether to include the cleaned and parsed HTML content of each result"},
+                            "include_images": {"type": "boolean", "description": "Whether to include images from search results"},
+                            "include_image_descriptions": {"type": "boolean", "description": "Whether to include descriptions with images"},
+                            "max_results": {"type": "integer", "description": "Maximum number of results to return (5-20)"},
+                            "async_search": {"type": "boolean", "description": "Whether to perform the search asynchronously"},
+                        },
+                        "required": ["what_is_your_intent", "query"],
+                    },
+                    "strict": False,
+                },
+            }
+        ]
+        # Mock mcp registered tool as a predefined schema for avoiding ci timeout
+        mock_fetch.return_value = tool_schema
         search_rollout_config.multi_turn.max_turns = 1
         rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         req = rollout._preprocess_prompt_to_async_rollout_requests(search_data_proto, n=1)[0]
@@ -209,13 +310,45 @@ class TestRolloutWithMCPSearchTools:
             tool_calls=None,
         )
 
+    @patch.object(MCPClientManager, "fetch_tool_schemas", return_value=None)
     @patch.object(MCPSearchTool, "execute", new_callable=AsyncMock)
     @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
     @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
     @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
-    def test_tool_call_basic_case(self, mock_sampling, mock_engine, mock_env, mock_execute, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
+    def test_tool_call_basic_case(self, mock_sampling, mock_engine, mock_env, mock_execute, mock_fetch, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
         _, expect_turn_array, tool_return_array = search_data
-
+        tool_schema = [
+            {
+                "type": "function",
+                "function": {
+                    "name": "tavily_search_tool",
+                    "description": "A powerful web search tool...",
+                    "parameters": {
+                        "type": "object",
+                        "properties": {
+                            "what_is_your_intent": {"type": "string", "description": "Describe your intent for using Tavily"},
+                            "query": {"type": "string", "description": "Search query"},
+                            "search_depth": {"type": "string", "description": "The depth of the search ('basic' or 'advanced')"},
+                            "topic": {"type": "string", "description": "The category of the search ('general' or 'news')"},
+                            "days": {"type": "integer", "description": "Number of days back to include in search results (only for 'news' topic)"},
+                            "time_range": {"type": "string", "description": "Time range for results ('day', 'week', 'month', 'year', 'd', 'w', 'm', 'y')"},
+                            "include_domains": {"type": "array", "description": "List of domains to specifically include in search results"},
+                            "exclude_domains": {"type": "array", "description": "List of domains to specifically exclude from search results"},
+                            "include_answer": {"type": "boolean", "description": "Whether to include an answer summary generated by an LLM"},
+                            "include_raw_content": {"type": "boolean", "description": "Whether to include the cleaned and parsed HTML content of each result"},
+                            "include_images": {"type": "boolean", "description": "Whether to include images from search results"},
+                            "include_image_descriptions": {"type": "boolean", "description": "Whether to include descriptions with images"},
+                            "max_results": {"type": "integer", "description": "Maximum number of results to return (5-20)"},
+                            "async_search": {"type": "boolean", "description": "Whether to perform the search asynchronously"},
+                        },
+                        "required": ["what_is_your_intent", "query"],
+                    },
+                    "strict": False,
+                },
+            }
+        ]
+        # Mock mcp registered tool as a predefined schema for avoiding ci timeout
+        mock_fetch.return_value = tool_schema
         # Mock search tool execution to return predefined responses
         mock_execute.side_effect = [(msg, 0.0, {"status": "success"}) for msg in tool_return_array]
 
@@ -256,13 +389,45 @@ class TestRolloutWithMCPSearchTools:
                 search_counter += 1
         assert search_counter == 2
 
+    @patch.object(MCPClientManager, "fetch_tool_schemas", return_value=None)
     @patch.object(MCPSearchTool, "execute", new_callable=AsyncMock)
     @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
     @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
     @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
-    def test_tool_call_batch_case(self, mock_sampling, mock_engine, mock_env, mock_execute, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
+    def test_tool_call_batch_case(self, mock_sampling, mock_engine, mock_env, mock_execute, mock_fetch, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
         _, expect_turn_array, tool_return_array = search_data
-
+        tool_schema = [
+            {
+                "type": "function",
+                "function": {
+                    "name": "tavily_search_tool",
+                    "description": "A powerful web search tool...",
+                    "parameters": {
+                        "type": "object",
+                        "properties": {
+                            "what_is_your_intent": {"type": "string", "description": "Describe your intent for using Tavily"},
+                            "query": {"type": "string", "description": "Search query"},
+                            "search_depth": {"type": "string", "description": "The depth of the search ('basic' or 'advanced')"},
+                            "topic": {"type": "string", "description": "The category of the search ('general' or 'news')"},
+                            "days": {"type": "integer", "description": "Number of days back to include in search results (only for 'news' topic)"},
+                            "time_range": {"type": "string", "description": "Time range for results ('day', 'week', 'month', 'year', 'd', 'w', 'm', 'y')"},
+                            "include_domains": {"type": "array", "description": "List of domains to specifically include in search results"},
+                            "exclude_domains": {"type": "array", "description": "List of domains to specifically exclude from search results"},
+                            "include_answer": {"type": "boolean", "description": "Whether to include an answer summary generated by an LLM"},
+                            "include_raw_content": {"type": "boolean", "description": "Whether to include the cleaned and parsed HTML content of each result"},
+                            "include_images": {"type": "boolean", "description": "Whether to include images from search results"},
+                            "include_image_descriptions": {"type": "boolean", "description": "Whether to include descriptions with images"},
+                            "max_results": {"type": "integer", "description": "Maximum number of results to return (5-20)"},
+                            "async_search": {"type": "boolean", "description": "Whether to perform the search asynchronously"},
+                        },
+                        "required": ["what_is_your_intent", "query"],
+                    },
+                    "strict": False,
+                },
+            }
+        ]
+        # Mock mcp registered tool as a predefined schema for avoiding ci timeout
+        mock_fetch.return_value = tool_schema
         # Mock tool execution for large batch (100 requests * 2 calls each)
         mock_execute.side_effect = [
             (tool_return_array[0], 0.0, {"status": "success"}),
